Move rollout in a different file
