+++ internal/model/flag_pub_test.go
@@ -777,48 +777,3 @@ func TestFlag_String(t *testing.T) {
 		})
 	}
 }
-
-func TestExperimentation_String(t *testing.T) {
-	type fields struct {
-		StartDate *time.Time
-		EndDate   *time.Time
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		want   string
-	}{
-		{
-			name: "both dates",
-			fields: fields{
-				StartDate: testutils.Time(time.Unix(1095379400, 0)),
-				EndDate:   testutils.Time(time.Unix(1095379500, 0)),
-			},
-			want: "start:[2004-09-17T00:03:20Z] end:[2004-09-17T00:05:00Z]",
-		},
-		{
-			name: "only start date",
-			fields: fields{
-				StartDate: testutils.Time(time.Unix(1095379400, 0)),
-			},
-			want: "start:[2004-09-17T00:03:20Z]",
-		},
-		{
-			name: "only end date",
-			fields: fields{
-				EndDate: testutils.Time(time.Unix(1095379500, 0)),
-			},
-			want: "end:[2004-09-17T00:05:00Z]",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			e := model.Experimentation{
-				StartDate: tt.fields.StartDate,
-				EndDate:   tt.fields.EndDate,
-			}
-			got := e.String()
-			assert.Equal(t, tt.want, got)
-		})
-	}
-}
